Add to cart feature
To make things easier, updated structure of users collection in firestore. The document ID is now the same as the UID in firebase authentication.

Implemented add to cart function in product details page

Creates new document in cart if the user does not currently have the item in their cart. Else, it adds the quantity of the product to their cart (qty = currentQty + addedQty)

Fixed wording in Register.vue